Divisors update, 021, update, made 132 better (but still not great)

+++ src/euler/PE_021.java
@@ -18,15 +18,15 @@ public class PE_021 {
     }
 
     private static int amicablePair(int n) {
-        long[] divisors = Divisors.divisors(n);
+        int[] divisors = Divisors.divisors(n);
         int sum = 0;
-        for (long divisor : divisors) sum += (int) divisor;
+        for (int divisor : divisors) sum += divisor;
         sum -= n;
         int otherNumber = sum;
         if (otherNumber < 1) return -1;
         divisors = Divisors.divisors(otherNumber);
         sum = 0;
-        for (long divisor : divisors) sum += (int) divisor;
+        for (int divisor : divisors) sum += divisor;
         sum -= otherNumber;
         return sum == n && otherNumber != n ? otherNumber : -1;
     }
